Cleanup code. Use clang on Linux.

+++ p_view.c
@@ -1561,7 +1561,7 @@ void ClientEndServerFrame(edict_t *ent)
 		bobtime *= 4;
 
 	bobcycle = (int)bobtime;
-	bobfracsin = fabsf(sin(bobtime*M_PI));
+	bobfracsin = fabs(sin(bobtime*M_PI));
 
 	// detect hitting the floor
 	P_FallingDamage(ent);
